feat(jwt): add more fixtures and move to conftest
